requirements.txt
python-telegram-bot==20.0
flask==2.3.2
gunicorn==20.1.0
python-dotenv==1.0.0
apscheduler==3.10.1  # Versão compatível com PTB 20.x
